Add restrictions in command
