+++ .github/workflows/linter-schemes.yml
@@ -1,6 +1,5 @@
 name: Format Json Schemes
 
-# Run this workflow every time a new commit pushed to your repository
 on: workflow_dispatch
 
 jobs:
