+++ .github/workflows/linter.yml
@@ -1,6 +1,5 @@
 name: Code Linter
 
-# Run this workflow every time a new commit pushed to your repository
 on: workflow_dispatch
 
 jobs:
